+++ src/main/java/net/ttk1/hideme/command/HideMeCommandExecutor.java
@@ -1,9 +1,11 @@
-package net.ttk1.command;
+package net.ttk1.hideme.command;
 
 import com.google.inject.Inject;
-import net.ttk1.HideMe;
-import net.ttk1.api.PlayerManager;
-import net.ttk1.command.subcommand.SubCommand;
+
+import net.ttk1.hideme.HideMe;
+import net.ttk1.hideme.api.PlayerManager;
+import net.ttk1.hideme.command.subcommand.SubCommand;
+
 import org.bukkit.command.Command;
 import org.bukkit.command.CommandExecutor;
 import org.bukkit.command.CommandSender;
