+++ src/main/java/net/ttk1/hideme/command/subcommand/VersionCommand.java
@@ -1,11 +1,11 @@
-package net.ttk1.command.subcommand;
+package net.ttk1.hideme.command.subcommand;
 
-import net.ttk1.HideMe;
-import net.ttk1.api.PlayerManager;
+import net.ttk1.hideme.HideMe;
+import net.ttk1.hideme.api.PlayerManager;
+
 import org.bukkit.command.CommandSender;
 
 import java.util.HashSet;
-import java.util.List;
 import java.util.Set;
 
 public class VersionCommand implements SubCommand {
